+++ lib/device/dev-md.c
@@ -322,8 +322,8 @@ static int _md_sysfs_attribute_snprintf(char *path, size_t size,
 	if (MAJOR(dev) != dt->md_major)
 		return ret;
 
-	ret = dm_snprintf(path, size, "%s/dev/block/%d:%d/md/%s", sysfs_dir,
-			  (int)MAJOR(dev), (int)MINOR(dev), attribute);
+	ret = dm_snprintf(path, size, "%s/dev/block/%u:%u/md/%s", sysfs_dir,
+			  MAJOR(dev), MINOR(dev), attribute);
 	if (ret < 0) {
 		log_error("dm_snprintf md %s failed", attribute);
 		return ret;
@@ -335,8 +335,8 @@ static int _md_sysfs_attribute_snprintf(char *path, size_t size,
 			return ret;
 		}
 		/* old sysfs structure */
-		ret = dm_snprintf(path, size, "%s/block/md%d/md/%s",
-				  sysfs_dir, (int)MINOR(dev), attribute);
+		ret = dm_snprintf(path, size, "%s/block/md%u/md/%s",
+				  sysfs_dir, MINOR(dev), attribute);
 		if (ret < 0) {
 			log_error("dm_snprintf old md %s failed", attribute);
 			return ret;
